+++ NEWS.md
@@ -4,6 +4,7 @@
 * Speed up 'unsetting' of large formulas/compounds objects, which affects eg plotting TP similarities, annotatedPeakList() etc (reported by Alessia Ore)
 * `report()`: correctly handle removed suspect hits while reporting TP similarities (reported by Alessia Ore)
 * Added visual clue to `checkFeatures()` and `checkComponents()` to see whether a feature or component is marked to be removed (pull request #117 as suggested by Leon Saal)
+* `loadMSLibrary()`: don't always set `Ion_mode` of records to positive and guess missing `Ion_mode` data (issue #119)
 
 
 # patRoon 2.3.3